Added client code to get albums from a GCP cloud function.
Better messages coming from the cover manager to the UI (errors, what is going on, progress bar).

Details:
Need to launch LoadBackupData() after creating a new instance of the cover manager. This lets us provide callbacks before the work is done. Could provide those in the constructor...

Needs the following packages:
NuGet\Install-Package Google.Apis.Auth -Version 1.62.1
NuGet\Install-Package System.Net.Http.Json -Version 7.0.1
There might be a way to add those directly in some conf file and skip the NuGet part entirely...

Messages are propagated with 3 events, one for when a new cover is found, one for the progress bar update, and one with a message. Using the new fancy .? way to invoke delegates.

Changed UI to directly launch the parsing on the return of the FolderBrowserDialog form. There's no actual need to display the folder in a textbox.